move fixtures into xml files and small refactor.
