--- tests/providers/s3/fixtures/fixtures.py
@@ -1,323 +1,0 @@
-import pytest
-
-from collections import OrderedDict
-
-from waterbutler.providers.s3.metadata import (
-    S3FileMetadataHeaders,
-    S3FileMetadata,
-    S3FolderMetadata,
-    S3FolderKeyMetadata,
-    S3Revision
-)
-
-
-@pytest.fixture
-def auth():
-    return {
-        'name': 'cat',
-        'email': 'cat@cat.com',
-    }
-
-
-@pytest.fixture
-def credentials():
-    return {
-        'access_key': 'Dont dead',
-        'secret_key': 'open inside',
-    }
-
-
-@pytest.fixture
-def settings():
-    return {
-        'bucket': 'that kerning',
-        'encrypt_uploads': False
-    }
-
-
-@pytest.fixture
-def file_content():
-    return b'sleepy'
-
-@pytest.fixture
-def folder_metadata():
-    return b'''<?xml version="1.0" encoding="UTF-8"?>
-        <ListBucketResult xmlns="http://s3.amazonaws.com/doc/2006-03-01/">
-            <Name>bucket</Name>
-            <Prefix/>
-            <Marker/>
-            <MaxKeys>1000</MaxKeys>
-            <IsTruncated>false</IsTruncated>
-            <Contents>
-                <Key>my-image.jpg</Key>
-                <LastModified>2009-10-12T17:50:30.000Z</LastModified>
-                <ETag>&quot;fba9dede5f27731c9771645a39863328&quot;</ETag>
-                <Size>434234</Size>
-                <StorageClass>STANDARD</StorageClass>
-                <Owner>
-                    <ID>75aa57f09aa0c8caeab4f8c24e99d10f8e7faeebf76c078efc7c6caea54ba06a</ID>
-                    <DisplayName>mtd@amazon.com</DisplayName>
-                </Owner>
-            </Contents>
-            <Contents>
-                <Key>my-third-image.jpg</Key>
-                <LastModified>2009-10-12T17:50:30.000Z</LastModified>
-                <ETag>&quot;1b2cf535f27731c974343645a3985328&quot;</ETag>
-                <Size>64994</Size>
-                <StorageClass>STANDARD</StorageClass>
-                <Owner>
-                    <ID>75aa57f09aa0c8caeab4f8c24e99d10f8e7faeebf76c078efc7c6caea54ba06a</ID>
-                    <DisplayName>mtd@amazon.com</DisplayName>
-                </Owner>
-            </Contents>
-            <CommonPrefixes>
-                <Prefix>   photos/</Prefix>
-            </CommonPrefixes>
-        </ListBucketResult>'''
-
-@pytest.fixture
-def folder_single_thing_metadata():
-    return b'''<?xml version="1.0" encoding="UTF-8"?>
-        <ListBucketResult xmlns="http://s3.amazonaws.com/doc/2006-03-01/">
-            <Name>bucket</Name>
-            <Prefix/>
-            <Marker/>
-            <MaxKeys>1000</MaxKeys>
-            <IsTruncated>false</IsTruncated>
-            <Contents>
-                <Key>my-image.jpg</Key>
-                <LastModified>2009-10-12T17:50:30.000Z</LastModified>
-                <ETag>&quot;fba9dede5f27731c9771645a39863328&quot;</ETag>
-                <Size>434234</Size>
-                <StorageClass>STANDARD</StorageClass>
-                <Owner>
-                    <ID>75aa57f09aa0c8caeab4f8c24e99d10f8e7faeebf76c078efc7c6caea54ba06a</ID>
-                    <DisplayName>mtd@amazon.com</DisplayName>
-                </Owner>
-            </Contents>
-            <CommonPrefixes>
-                <Prefix>   photos/</Prefix>
-            </CommonPrefixes>
-        </ListBucketResult>'''
-
-
-@pytest.fixture
-def just_a_folder_metadata():
-    return b'''<?xml version="1.0" encoding="UTF-8"?>
-        <ListBucketResult xmlns="http://s3.amazonaws.com/doc/2006-03-01/">
-            <Name>bucket</Name>
-            <Prefix/>
-            <Marker/>
-            <MaxKeys>1000</MaxKeys>
-            <IsTruncated>false</IsTruncated>
-            <Contents>
-                <Key>naptime/</Key>
-                <LastModified>2009-10-12T17:50:30.000Z</LastModified>
-                <ETag>&quot;fba9dede5f27731c9771645a39863328&quot;</ETag>
-                <Size>0</Size>
-                <StorageClass>STANDARD</StorageClass>
-                <Owner>
-                    <ID>75aa57f09aa0c8caeab4f8c24e99d10f8e7faeebf76c078efc7c6caea54ba06a</ID>
-                    <DisplayName>mtd@amazon.com</DisplayName>
-                </Owner>
-            </Contents>
-        </ListBucketResult>'''
-
-
-@pytest.fixture
-def contents_and_self():
-    return b'''<?xml version="1.0" encoding="UTF-8"?>
-        <ListBucketResult xmlns="http://s3.amazonaws.com/doc/2006-03-01/">
-            <Name>bucket</Name>
-            <Prefix/>
-            <Marker/>
-            <MaxKeys>1000</MaxKeys>
-            <IsTruncated>false</IsTruncated>
-            <Contents>
-                <Key>thisfolder/</Key>
-                <LastModified>2009-10-12T17:50:30.000Z</LastModified>
-                <ETag>&quot;fba9dede5f27731c9771645a39863328&quot;</ETag>
-                <Size>0</Size>
-                <StorageClass>STANDARD</StorageClass>
-                <Owner>
-                    <ID>75aa57f09aa0c8caeab4f8c24e99d10f8e7faeebf76c078efc7c6caea54ba06a</ID>
-                    <DisplayName>mtd@amazon.com</DisplayName>
-                </Owner>
-            </Contents>
-            <Contents>
-                <Key>thisfolder/item1</Key>
-                <LastModified>2009-10-12T17:50:30.000Z</LastModified>
-                <ETag>&quot;fba9dede5f27731c9771645a39863328&quot;</ETag>
-                <Size>0</Size>
-                <StorageClass>STANDARD</StorageClass>
-                <Owner>
-                    <ID>75aa57f09aa0c8caeab4f8c24e99d10f8e7faeebf76c078efc7c6caea54ba06a</ID>
-                    <DisplayName>mtd@amazon.com</DisplayName>
-                </Owner>
-            </Contents>
-            <Contents>
-                <Key>thisfolder/item2</Key>
-                <LastModified>2009-10-12T17:50:30.000Z</LastModified>
-                <ETag>&quot;fba9dede5f27731c9771645a39863328&quot;</ETag>
-                <Size>0</Size>
-                <StorageClass>STANDARD</StorageClass>
-                <Owner>
-                    <ID>75aa57f09aa0c8caeab4f8c24e99d10f8e7faeebf76c078efc7c6caea54ba06a</ID>
-                    <DisplayName>mtd@amazon.com</DisplayName>
-                </Owner>
-            </Contents>
-        </ListBucketResult>'''
-
-
-@pytest.fixture
-def folder_empty_metadata():
-    return b'''<?xml version="1.0" encoding="UTF-8"?>
-        <ListBucketResult xmlns="http://s3.amazonaws.com/doc/2006-03-01/">
-            <Name>bucket</Name>
-            <Prefix/>
-            <Marker/>
-            <MaxKeys>1000</MaxKeys>
-            <IsTruncated>false</IsTruncated>
-        </ListBucketResult>'''
-
-
-@pytest.fixture
-def file_header_metadata():
-    return {
-        'CONTENT-LENGTH': 9001,
-        'LAST-MODIFIED': 'SomeTime',
-        'CONTENT-TYPE': 'binary/octet-stream',
-        'ETAG': '"fba9dede5f27731c9771645a39863328"',
-        'X-AMZ-SERVER-SIDE-ENCRYPTION': 'AES256'
-    }
-
-
-@pytest.fixture
-def version_metadata():
-    return b'''<?xml version="1.0" encoding="UTF-8"?>
-
-    <ListVersionsResult xmlns="http://s3.amazonaws.com/doc/2006-03-01">
-        <Name>bucket</Name>
-        <Prefix>my</Prefix>
-        <KeyMarker/>
-        <VersionIdMarker/>
-        <MaxKeys>5</MaxKeys>
-        <IsTruncated>false</IsTruncated>
-        <Version>
-            <Key>my-image.jpg</Key>
-            <VersionId>3/L4kqtJl40Nr8X8gdRQBpUMLUo</VersionId>
-            <IsLatest>true</IsLatest>
-            <LastModified>2009-10-12T17:50:30.000Z</LastModified>
-            <ETag>&quot;fba9dede5f27731c9771645a39863328&quot;</ETag>
-            <Size>434234</Size>
-            <StorageClass>STANDARD</StorageClass>
-            <Owner>
-                <ID>75aa57f09aa0c8caeab4f8c24e99d10f8e7faeebf76c078efc7c6caea54ba06a</ID>
-                <DisplayName>mtd@amazon.com</DisplayName>
-            </Owner>
-        </Version>
-        <Version>
-            <Key>my-image.jpg</Key>
-            <VersionId>QUpfdndhfd8438MNFDN93jdnJFkdmqnh893</VersionId>
-            <IsLatest>false</IsLatest>
-            <LastModified>2009-10-10T17:50:30.000Z</LastModified>
-            <ETag>&quot;9b2cf535f27731c974343645a3985328&quot;</ETag>
-            <Size>166434</Size>
-            <StorageClass>STANDARD</StorageClass>
-            <Owner>
-                <ID>75aa57f09aa0c8caeab4f8c24e99d10f8e7faeebf76c078efc7c6caea54ba06a</ID>
-                <DisplayName>mtd@amazon.com</DisplayName>
-            </Owner>
-        </Version>
-        <Version>
-            <Key>my-image.jpg</Key>
-            <VersionId>UIORUnfndfhnw89493jJFJ</VersionId>
-            <IsLatest>false</IsLatest>
-            <LastModified>2009-10-11T12:50:30.000Z</LastModified>
-            <ETag>&quot;772cf535f27731c974343645a3985328&quot;</ETag>
-            <Size>64</Size>
-            <StorageClass>STANDARD</StorageClass>
-            <Owner>
-                <ID>75aa57f09aa0c8caeab4f8c24e99d10f8e7faeebf76c078efc7c6caea54ba06a</ID>
-                <DisplayName>mtd@amazon.com</DisplayName>
-            </Owner>
-        </Version>
-    </ListVersionsResult>'''
-
-
-@pytest.fixture
-def single_version_metadata():
-    return b'''<?xml version="1.0" encoding="UTF-8"?>
-
-    <ListVersionsResult xmlns="http://s3.amazonaws.com/doc/2006-03-01">
-        <Name>bucket</Name>
-        <Prefix>my</Prefix>
-        <KeyMarker/>
-        <VersionIdMarker/>
-        <MaxKeys>5</MaxKeys>
-        <IsTruncated>false</IsTruncated>
-        <Version>
-            <Key>single-version.file</Key>
-            <VersionId>3/L4kqtJl40Nr8X8gdRQBpUMLUo</VersionId>
-            <IsLatest>true</IsLatest>
-            <LastModified>2009-10-12T17:50:30.000Z</LastModified>
-            <ETag>&quot;fba9dede5f27731c9771645a39863328&quot;</ETag>
-            <Size>434234</Size>
-            <StorageClass>STANDARD</StorageClass>
-            <Owner>
-                <ID>75aa57f09aa0c8caeab4f8c24e99d10f8e7faeebf76c078efc7c6caea54ba06a</ID>
-                <DisplayName>mtd@amazon.com</DisplayName>
-            </Owner>
-        </Version>
-    </ListVersionsResult>'''
-
-
-@pytest.fixture
-def file_metadata_headers_object(file_header_metadata):
-    return S3FileMetadataHeaders('/test-path', file_header_metadata)
-
-
-@pytest.fixture
-def file_metadata_object():
-    content = OrderedDict(Key='my-image.jpg',
-                          LastModified='2009-10-12T17:50:30.000Z',
-                          ETag="fba9dede5f27731c9771645a39863328",
-                          Size='434234',
-                          StorageClass='STANDARD')
-
-    return S3FileMetadata(content)
-
-
-@pytest.fixture
-def folder_key_metadata_object():
-    content = OrderedDict(Key='naptime/',
-                          LastModified='2009-10-12T17:50:30.000Z',
-                          ETag='"fba9dede5f27731c9771645a39863328"',
-                          Size='0',
-                          StorageClass='STANDARD')
-
-    return S3FolderKeyMetadata(content)
-
-
-@pytest.fixture
-def folder_metadata_object():
-    content = OrderedDict(Prefix='photos/')
-    return S3FolderMetadata(content)
-
-
-@pytest.fixture
-def revision_metadata_object():
-    content = OrderedDict(Key='single-version.file',
-                VersionId='3/L4kqtJl40Nr8X8gdRQBpUMLUo',
-                IsLatest='true',
-                LastModified='2009-10-12T17:50:30.000Z',
-                ETag='"fba9dede5f27731c9771645a39863328"',
-                Size=434234,
-                StorageClass='STANDARD',
-                Owner=OrderedDict(ID='75aa57f09aa0c8caeab4f8c24e99d10f8e7faeebf76c078efc7c6caea54ba06a',
-                                  DisplayName='mtd@amazon.com')
-                )
-
-    return S3Revision(content)
-
-
