+++ tests/providers/s3/fixtures/folder_metadata.xml
@@ -1,0 +1,34 @@
+<?xml version="1.0" encoding="UTF-8"?>
+
+<ListBucketResult xmlns="http://s3.amazonaws.com/doc/2006-03-01/">
+    <Name>bucket</Name>
+    <Prefix/>
+    <Marker/>
+    <MaxKeys>1000</MaxKeys>
+    <IsTruncated>false</IsTruncated>
+    <Contents>
+        <Key>my-image.jpg</Key>
+        <LastModified>2009-10-12T17:50:30.000Z</LastModified>
+        <ETag>&quot;fba9dede5f27731c9771645a39863328&quot;</ETag>
+        <Size>434234</Size>
+        <StorageClass>STANDARD</StorageClass>
+        <Owner>
+            <ID>75aa57f09aa0c8caeab4f8c24e99d10f8e7faeebf76c078efc7c6caea54ba06a</ID>
+            <DisplayName>mtd@amazon.com</DisplayName>
+        </Owner>
+    </Contents>
+    <Contents>
+        <Key>my-third-image.jpg</Key>
+        <LastModified>2009-10-12T17:50:30.000Z</LastModified>
+        <ETag>&quot;1b2cf535f27731c974343645a3985328&quot;</ETag>
+        <Size>64994</Size>
+        <StorageClass>STANDARD</StorageClass>
+        <Owner>
+            <ID>75aa57f09aa0c8caeab4f8c24e99d10f8e7faeebf76c078efc7c6caea54ba06a</ID>
+            <DisplayName>mtd@amazon.com</DisplayName>
+        </Owner>
+    </Contents>
+    <CommonPrefixes>
+        <Prefix>   photos/</Prefix>
+    </CommonPrefixes>
+</ListBucketResult>
