+++ tests/providers/s3/fixtures/version_metadata.xml
@@ -1,0 +1,49 @@
+<?xml version="1.0" encoding="UTF-8"?>
+
+<ListVersionsResult xmlns="http://s3.amazonaws.com/doc/2006-03-01">
+    <Name>bucket</Name>
+    <Prefix>my</Prefix>
+    <KeyMarker/>
+    <VersionIdMarker/>
+    <MaxKeys>5</MaxKeys>
+    <IsTruncated>false</IsTruncated>
+    <Version>
+        <Key>my-image.jpg</Key>
+        <VersionId>3/L4kqtJl40Nr8X8gdRQBpUMLUo</VersionId>
+        <IsLatest>true</IsLatest>
+        <LastModified>2009-10-12T17:50:30.000Z</LastModified>
+        <ETag>&quot;fba9dede5f27731c9771645a39863328&quot;</ETag>
+        <Size>434234</Size>
+        <StorageClass>STANDARD</StorageClass>
+        <Owner>
+            <ID>75aa57f09aa0c8caeab4f8c24e99d10f8e7faeebf76c078efc7c6caea54ba06a</ID>
+            <DisplayName>mtd@amazon.com</DisplayName>
+        </Owner>
+    </Version>
+    <Version>
+        <Key>my-image.jpg</Key>
+        <VersionId>QUpfdndhfd8438MNFDN93jdnJFkdmqnh893</VersionId>
+        <IsLatest>false</IsLatest>
+        <LastModified>2009-10-10T17:50:30.000Z</LastModified>
+        <ETag>&quot;9b2cf535f27731c974343645a3985328&quot;</ETag>
+        <Size>166434</Size>
+        <StorageClass>STANDARD</StorageClass>
+        <Owner>
+            <ID>75aa57f09aa0c8caeab4f8c24e99d10f8e7faeebf76c078efc7c6caea54ba06a</ID>
+            <DisplayName>mtd@amazon.com</DisplayName>
+        </Owner>
+    </Version>
+    <Version>
+        <Key>my-image.jpg</Key>
+        <VersionId>UIORUnfndfhnw89493jJFJ</VersionId>
+        <IsLatest>false</IsLatest>
+        <LastModified>2009-10-11T12:50:30.000Z</LastModified>
+        <ETag>&quot;772cf535f27731c974343645a3985328&quot;</ETag>
+        <Size>64</Size>
+        <StorageClass>STANDARD</StorageClass>
+        <Owner>
+            <ID>75aa57f09aa0c8caeab4f8c24e99d10f8e7faeebf76c078efc7c6caea54ba06a</ID>
+            <DisplayName>mtd@amazon.com</DisplayName>
+        </Owner>
+    </Version>
+</ListVersionsResult>
